// src/components/apis/modules/fast-project/fast-project-api-types.ts
export type CreateFastProjectRequest = {
  typeSourceId: number;
  sourceDetailId: number;
  operationTypeId: number;
  usageLogementId: number;
  etatProjetId: number;
  firstName: string;
  lastName: string;
  civilite: string;
  email: string;
  mobilePhone: string;
  mobilePhoneIso2: string;
  place: string;
  customerType: string;
};
export type CreateFastProjectRequestResponse = {
  statusCode: number;
  message: string;
  data: {
    id: string;
    userId: string;
    sourceDetailId: string;
    operationTypeId: string;
    originId: string | null;
    durationInYears: number;
    isFakeApplication: boolean;
    isDuplicateApplication: boolean;
    isFromForm: boolean;
    isLocked: boolean;
    isCompanyCredit: boolean;
    isAccountKeepingVerified: boolean;
    withBien: boolean;
    clientUserId: string | null;
    apporteurId: string | null;
    backofficeId: string | null;
    commercialeId: string | null;
    assistanteId: string | null;
    ownerId: string | null;
    preteurId: string | null;
    preferencesId: string | null;
    typeOperationAssuranceId: string | null;
    createdAt: string;
    updatedAt: string;
    deletedAt: string | null;
    applicationInstructionId: string | null;
    typeSourceId: string;
    property: {
      id: string;
      applicationId: string;
      etatLogementId: string;
      proprieteBienId: string | null;
      place: string;
      hasSamePropInFrench: boolean;
      nbrPiece: number | null;
      surfaceHabitable: number | null;
      surfaceTerrain: number | null;
      dpe: string | null;
      futureOccupants: number | null;
      usageLogementId: string;
      numDepart: string | null;
      codeInsee: string | null;
      refVendeur: string | null;
      refNumReservation: string | null;
      refNumProg: string | null;
      refNumLot: string | null;
      refAnnexes: string | null;
      refCadastre: string | null;
      isAccessionPsla: boolean;
      isRevenuteHlm: boolean;
      isTravLourdRemiseNeufPtz: boolean;
      isTravTransfLogementPtz: boolean;
      normeConstrucId: string | null;
      perfEneget: string | null;
      ageLogement: number | null;
      surfaceAnnex: number | null;
      surfaceCreatedTrav: number | null;
      loceauxPro: boolean;
      zonePtz: string | null;
      zoneApl: string | null;
      isPrimoAccessionHp: boolean;
      taxeFonciere: number | null;
      revenuLocatif: number | null;
      cpEtr: string | null;
      cityEtr: string | null;
      souhaiteEstimationGratuite: boolean;
      createdAt: string;
      updatedAt: string;
      deletedAt: string | null;
    };
    applicationCustomers: {
      applicationId: string;
      customerId: string;
      customerType: string;
      customer: {
        id: string;
        firstName: string;
        lastName: string;
        sex: string | null;
        civilite: string;
        email: string;
        birthday: string | null;
        maritalStatus: string | null;
        matrimonialRegimes: string | null;
        weddingDate: string | null;
        weddingContractDate: string | null;
        divorceDate: string | null;
        nomJeuneFille: string | null;
        parentFirstRelation: string | null;
        menageWith: string | null;
        isEmpOrCoemp: boolean;
        isSci: boolean;
        sciPercent: number | null;
        isCaution: boolean;
        isSolidaire: boolean;
        isHypotecaire: boolean;
        diplomeGrandEcole: string | null;
        codeParrainage: string | null;
        codeParrain: string | null;
        locationFilledFromAvis: boolean;
        cniRectoScannedOcr: boolean;
        budgetInsightLink: string | null;
        userId: string | null;
        birthPlaceId: string | null;
        weddingPlaceId: string | null;
        countryLivingId: string | null;
        nationalityId: string | null;
        createdAt: string;
        updatedAt: string;
        deletedAt: string | null;
        customerPhoneNumbers: {
          customerId: string;
          phoneNumberId: string;
          phoneNumber: {
            id: string;
            number: string;
            category: string;
            countryId: string;
            createdAt: string;
            updatedAt: string;
            etablissementBancaireId: string | null;
            country: {
              iso2: string;
              iso3: string;
              numericCode: string;
              phoneCode: string;
              name: string;
              officialName: string;
              latitude: number;
              longitude: number;
              region: string;
              subregion: string;
              timezones: string[];
              currencies: string[];
              emoji: string;
              emojiU: string | null;
              tld: string;
              createdAt: string;
              updatedAt: string;
            };
          };
        }[];
      };
    }[];
  };
};